% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mym2r.R
\name{prepare.r2mym}
\alias{prepare.r2mym}
\title{prepare.r2mym helps you merge an R table with the data.frame output you want for MyM}
\usage{
prepare.r2mym(data, stub, value.var, NaN.value = -9999)
}
\arguments{
\item{data}{Your data frame with all cols}

\item{stub}{Your preferred subset of DIMs}

\item{value.var}{Indication of  your value var}

\item{NaN.value}{Preset to -9999}
}
\value{
Returns smaller dataset that can be inserted into write.r2mym()
}
\description{
prepare.r2mym helps you merge an R table with the data.frame output you want for MyM
}
