% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mym2r.R
\name{read.mym2r.dims}
\alias{read.mym2r.dims}
\title{read.mym2r.dims: function to extract only the dimensions from the header in a mym file}
\usage{
read.mym2r.dims(path.to.mym.file)
}
\arguments{
\item{path.to.mym.file}{Reference to the location of the MyM file}
}
\value{
vector of dimensions as listed in the header / first line of the mym file
}
\description{
read.mym2r.dims: function to extract only the dimensions from the header in a mym file
}
