% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mym2r.R
\name{read.mym2r.nice}
\alias{read.mym2r.nice}
\title{read.mym2r.nice: Transform a MyM files into a readable datafile in R}
\usage{
read.mym2r.nice(mym.folder = "", scen.econ = "", filename = NULL,
  varname = NULL, collist = NULL, namecols = NULL,
  yearheader = "year", yearsrun = 1971:2100, novarname = FALSE)
}
\arguments{
\item{mym.folder}{"input/water model/mym output/"   # note: the exact file path would be: "input/water model/mym output/SSP2/ElecWWD_3_Int.out"}

\item{varname}{Provide the varname in the TIMER model here}

\item{yearheader}{The title of your "time" vector}

\item{yearsrun}{Provide indication here of your time sequence (defaults to 1971:2100)}

\item{novarname}{FALSE if there is a name for your variable}
}
\value{
produces a MyM data file with your own defined file extension
}
\description{
read.mym2r.nice: Transform a MyM files into a readable datafile in R
}
