% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mym2r.R
\name{read.mym2r}
\alias{read.mym2r}
\title{read.mym2r}
\usage{
read.mym2r(path.to.mym.file, yearheader = "year", yearsrun = NULL)
}
\arguments{
\item{path.to.mym.file}{path to MyM output file}

\item{yearheader}{defaults to 'year'}

\item{yearsrun}{defaults to NULL}
}
\value{
R dataframe in 'long' format (many rows, each dimension as a 'factor' in its column, and only one column with actual values). Requires path to MyM output file
}
\description{
read.mym2r
}
