% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mym2r.R
\name{read.mym2r.varname}
\alias{read.mym2r.varname}
\title{read.mym2r.varname}
\usage{
read.mym2r.varname(path.to.mym.file)
}
\arguments{
\item{path.to.mym.file}{vector of dimensions as listed in the header / first line of the mym file}
}
\value{
vector of dimensions as listed in the header / first line of the mym file
}
\description{
read.mym2r.varname
}
