% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mym2r.R
\name{write.r2mym}
\alias{write.r2mym}
\title{write.r2mym: your function to transform your R code into a MyM file that can be read by the TIMER model}
\usage{
write.r2mym(data, outputfile, value.var, MyM.vartype, MyM.varname,
  time.dependent = TRUE, sep.years = ",", sep.data = ",\\t",
  matrix.format = TRUE, sprintf.format = NULL, comment.line = NULL,
  header.line = NULL)
}
\arguments{
\item{data}{your R dataframe that needs to be transformed to MyM, should include only the data to be written (no excessive columns)}

\item{outputfile}{Your data output file, includes full path (if necessary) and file extension (e.g. ".dat")}

\item{value.var}{string indicating the column containing the numeric values}

\item{MyM.vartype}{the MyM variable type, e.g. 'REAL'}

\item{MyM.varname}{the full name for the variable in MyM, e.g. 'main.submodel.childunderweight_deaths'}

\item{time.dependent}{TRUE or FALSE, whether the variable is time dependent and the first column of data indicates the year}

\item{sep.years}{what characters to print after each year. '\n' is added automatically (Default is ',')}

\item{sep.data}{what characters to print after each regular number}

\item{matrix.format}{Default output is all the numbers for year X on one row. This setting makes multiple rows, leaving only the last dimension on one row (e.g. region) (matters for IMAGE input files!) (makes it slow, though)}

\item{comment.line}{Documentation line (e.g. original data source). It's an optional comment line to add at the top, crucial for IMAGE input files.}

\item{header.line}{optional, specifiy the exact header line yourself (replaces MyM.vartype, MyM.varname) - beware: more scope for human error!}
}
\value{
returns a data file in the path provided under outputfile
}
\description{
write.r2mym: your function to transform your R code into a MyM file that can be read by the TIMER model
}
